feat: implement shell command parser with quote handling
- Add a new parser for shell commands that supports quoting.
- Move command parsing logic from the shell worker to a dedicated parser module.
- Enhance command execution error handling with structured error information.
- Update documentation for command implementations to include usage examples.
