+++ apps/moo-dang/src/shell/commands.ts
@@ -2,16 +2,44 @@
  * Shell commands implementation for the virtual shell environment.
  *
  * Provides built-in commands using function-based architecture for better
- * testability and composability.
+ * testability and composability. All commands follow consistent patterns
+ * for error handling, result formatting, and execution timing.
  */
 
 import type {CommandExecutionResult, ExecutionContext, ShellCommand, VirtualFileSystem} from './types'
 
 /**
+ * Shell command execution error for command-specific failures.
+ *
+ * Provides structured error information for failed command executions,
+ * including the command name and appropriate exit code.
+ */
+export class ShellCommandError extends Error {
+  readonly command: string
+  readonly exitCode: number
+
+  constructor(command: string, exitCode: number, message: string) {
+    super(`${command}: ${message}`)
+    this.name = 'ShellCommandError'
+    this.command = command
+    this.exitCode = exitCode
+  }
+}
+
+/**
  * Create command execution result with timing information.
  *
  * Standardizes result format across all commands with consistent timing
- * and process tracking.
+ * and process tracking. This ensures all commands return results in a
+ * uniform structure for proper shell integration.
+ *
+ * @param context - Execution context containing process ID and environment
+ * @param command - Full command string that was executed
+ * @param stdout - Standard output from the command execution
+ * @param stderr - Standard error output from the command execution
+ * @param exitCode - Command exit code (0 for success, non-zero for error)
+ * @param startTime - Command start time for execution duration calculation
+ * @returns Structured command execution result
  */
 function createCommandResult(
   context: ExecutionContext,
@@ -32,7 +60,13 @@ function createCommandResult(
 }
 
 /**
- * Echo command - displays arguments as output.
+ * Create echo command implementation.
+ *
+ * Echo command displays its arguments as a line of text to standard output.
+ * This is one of the most fundamental shell commands, used for displaying
+ * text and testing command execution.
+ *
+ * @returns Shell command implementation for echo
  */
 function createEchoCommand(): ShellCommand {
   return {
@@ -47,7 +81,13 @@ function createEchoCommand(): ShellCommand {
 }
 
 /**
- * Pwd command - prints current working directory.
+ * Create pwd (print working directory) command implementation.
+ *
+ * Pwd command prints the absolute path of the current working directory.
+ * This command ignores all arguments and always returns the current directory
+ * from the execution context.
+ *
+ * @returns Shell command implementation for pwd
  */
 function createPwdCommand(): ShellCommand {
   return {
@@ -61,16 +101,35 @@ function createPwdCommand(): ShellCommand {
   }
 }
 
+/**
+ * File system entry metadata for directory listings.
+ *
+ * Represents a file or directory with Unix-like attributes for display
+ * in long-format directory listings (ls -l style).
+ */
 interface FileEntry {
-  name: string
-  type: 'file' | 'directory'
-  permissions: string
-  size: number
-  modified: string
+  /** File or directory name */
+  readonly name: string
+  /** Entry type (file or directory) */
+  readonly type: 'file' | 'directory'
+  /** Unix-style permission string (e.g., '-rw-r--r--') */
+  readonly permissions: string
+  /** File size in bytes */
+  readonly size: number
+  /** Last modified date string */
+  readonly modified: string
 }
 
 /**
- * Get detailed file listing with metadata.
+ * Get detailed file listing with metadata for long-format display.
+ *
+ * Transforms basic file names into detailed file entries with Unix-like
+ * metadata including permissions, size, and modification time. Uses
+ * heuristics to determine file vs directory types in the virtual environment.
+ *
+ * @param fileSystem - Virtual file system to query
+ * @param path - Directory path to list
+ * @returns Array of file entries with detailed metadata
  */
 async function getDetailedListing(fileSystem: VirtualFileSystem, path: string): Promise<FileEntry[]> {
   const entries = await fileSystem.listDirectory(path)
@@ -90,6 +149,13 @@ async function getDetailedListing(fileSystem: VirtualFileSystem, path: string): 
 
 /**
  * Format directory listing in long format (ls -l style).
+ *
+ * Creates Unix-like long-format directory listing with file permissions,
+ * sizes, dates, and names. Optionally filters hidden files (those starting with '.').
+ *
+ * @param entries - File entries to format
+ * @param showAll - Whether to show hidden files (starting with '.')
+ * @returns Formatted directory listing string
  */
 function formatLongListing(entries: FileEntry[], showAll: boolean): string {
   const filteredEntries = showAll ? entries : entries.filter(entry => !entry.name.startsWith('.'))
@@ -108,7 +174,14 @@ function formatLongListing(entries: FileEntry[], showAll: boolean): string {
 }
 
 /**
- * Ls command - lists directory contents with optional formatting.
+ * Create ls (list directory) command implementation.
+ *
+ * Lists directory contents with support for common flags like -l (long format)
+ * and -a (show all files including hidden). Handles both simple and detailed
+ * directory listings based on command flags.
+ *
+ * @param fileSystem - Virtual file system to query
+ * @returns Shell command implementation for ls
  */
 function createLsCommand(fileSystem: VirtualFileSystem): ShellCommand {
   return {
@@ -143,7 +216,14 @@ function createLsCommand(fileSystem: VirtualFileSystem): ShellCommand {
 }
 
 /**
- * Resolve relative paths to absolute paths.
+ * Resolve relative paths to absolute paths within the virtual file system.
+ *
+ * Converts relative path references to absolute paths by combining them
+ * with the current working directory. Handles edge cases like root directory.
+ *
+ * @param path - Path to resolve (can be absolute or relative)
+ * @param workingDirectory - Current working directory for relative resolution
+ * @returns Absolute path string
  */
 function resolvePath(path: string, workingDirectory: string): string {
   if (path.startsWith('/')) {
@@ -153,7 +233,14 @@ function resolvePath(path: string, workingDirectory: string): string {
 }
 
 /**
- * Cat command - displays file contents.
+ * Create cat (concatenate) command implementation.
+ *
+ * Cat command displays the contents of one or more files to standard output.
+ * It can handle multiple files and will concatenate their contents in order.
+ * Returns appropriate error messages for missing files or operands.
+ *
+ * @param fileSystem - Virtual file system to read files from
+ * @returns Shell command implementation for cat
  */
 function createCatCommand(fileSystem: VirtualFileSystem): ShellCommand {
   return {
@@ -198,7 +285,14 @@ function createCatCommand(fileSystem: VirtualFileSystem): ShellCommand {
 }
 
 /**
- * Help command - displays command information.
+ * Create help command implementation.
+ *
+ * Help command provides information about available commands and their usage.
+ * Can display general help listing all commands or specific help for individual commands.
+ * Integrates with the command registry to provide accurate, up-to-date information.
+ *
+ * @param commands - Map of available commands for help information
+ * @returns Shell command implementation for help
  */
 function createHelpCommand(commands: Map<string, ShellCommand>): ShellCommand {
   return {
@@ -247,7 +341,13 @@ function createHelpCommand(commands: Map<string, ShellCommand>): ShellCommand {
 }
 
 /**
- * Clear command - clears terminal screen.
+ * Create clear command implementation.
+ *
+ * Clear command clears the terminal screen using ANSI escape sequences.
+ * Sends escape codes to clear the entire screen and move cursor to home position.
+ * This command ignores all arguments.
+ *
+ * @returns Shell command implementation for clear
  */
 function createClearCommand(): ShellCommand {
   return {
@@ -261,10 +361,14 @@ function createClearCommand(): ShellCommand {
 }
 
 /**
- * Create standard set of shell commands.
+ * Create standard set of shell commands with file system integration.
  *
  * Initializes all built-in commands with proper dependencies and returns
- * a map for command lookup by name.
+ * a map for efficient command lookup by name. Each command is configured
+ * with access to the virtual file system where appropriate.
+ *
+ * @param fileSystem - Virtual file system instance for file operations
+ * @returns Map of command names to command implementations
  */
 export function createStandardCommands(fileSystem: VirtualFileSystem): Map<string, ShellCommand> {
   const commands = new Map<string, ShellCommand>()
