+++ apps/moo-dang/src/workers/shell.worker.ts
@@ -20,6 +20,7 @@ import type {
 import {consola} from 'consola'
 import {createStandardCommands} from '../shell/commands'
 import {ShellEnvironment} from '../shell/environment'
+import {parseCommand} from '../shell/parser'
 import {VirtualFileSystemImpl} from '../shell/virtual-file-system'
 
 interface ShellWorkerState {
@@ -249,44 +250,6 @@ function handleListProcesses(state: ShellWorkerState, _request: ListProcessesReq
     type: 'process-list',
     processes,
   }
-}
-
-/**
- * Parse command string into command and arguments with quote handling.
- *
- * Supports basic shell quoting with single and double quotes to handle arguments
- * containing spaces or special characters.
- */
-function parseCommand(command: string): string[] {
-  const parts: string[] = []
-  let current = ''
-  let inQuotes = false
-  let quoteChar = ''
-
-  const chars = Array.from(command)
-
-  for (const char of chars) {
-    if ((char === '"' || char === "'") && !inQuotes) {
-      inQuotes = true
-      quoteChar = char
-    } else if (char === quoteChar && inQuotes) {
-      inQuotes = false
-      quoteChar = ''
-    } else if (char === ' ' && !inQuotes) {
-      if (current) {
-        parts.push(current)
-        current = ''
-      }
-    } else {
-      current += char
-    }
-  }
-
-  if (current) {
-    parts.push(current)
-  }
-
-  return parts
 }
 
 /**
